Fix Pdf Upload
Adjust pdf upload component to the new redux dialog structure
